Client security - xss protection

+++ app/Http/Controllers/Client/LabelController.php
@@ -3,61 +3,72 @@
 namespace App\Http\Controllers\Client;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Client\Conversations\CreateLabelFormRequest;
+use App\Http\Requests\Client\Conversations\UpdateLabelFormRequest;
 use App\Models\Label;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use ProtoneMedia\LaravelXssProtection\Middleware\XssCleanInput;
 
 class LabelController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware(XssCleanInput::class)->only(['store', 'update']);
+    }
+
     public function index()
     {
         $labels = auth('client')->user()->labels;
 
         return view('client.labels.index')->with(compact('labels'));
     }
- 
-    public function store(Request $request)
+
+    public function store(CreateLabelFormRequest $request)
     {
-        try 
-        {
+        DB::beginTransaction();
+
+        try {
             // Label::create($request->all());
-            auth('client')->user()->labels()->create($request->all());
-            return redirect()->back()->with('success', 'Label was successfully created.');     
-        }
-        catch(Exception $e)
-        {
-            dd($e->getMessage());
+            auth('client')->user()->labels()->create($request->validated());
+            DB::commit();
+            return redirect()->back()->with('success', 'Label was successfully created.');
+        } catch (Exception $e) {
+            DB::rollBack();
+            return redirect()->back()->withErrors(['message' => 'Error: ' . $e->getMessage()]);
         }
     }
 
-    public function update(Label $label,Request $request)
+    public function update(Label $label, UpdateLabelFormRequest $request)
     {
-        try 
-        {
-            // Label::create($request->all());
-            // auth('client')->user()->labels()->create($request->all());
-            $label->update($request->all());
-            return redirect()->back()->with('success', 'Label was successfully updated.');     
-        }
-        catch(Exception $e)
-        {
-            dd($e->getMessage());
+        DB::beginTransaction();
+        try {
+            $label->update($request->validated());
+            DB::commit();
+            return redirect()->back()->with('success', 'Label was successfully updated.');
+        } catch (Exception $e) {
+            DB::rollBack();
+            return redirect()->back()->withErrors(['message' => 'Error: ' . $e->getMessage()]);
         }
     }
 
-    public function destroy(Label $label, Request $request)
+    public function destroy(Label $label)
     {
-        try
-        {              
+        DB::beginTransaction();
+        try {
+            if (!$label->isOwned()) {
+                return redirect()->back()->withErrors(['message' => 'Error: Unauthorized Operation!']);
+            }
+
             $label->delete();
 
+            DB::commit();
             return redirect(route('client.labels.index'))->with('success', 'Label was deleted successfully!');
-        }
-        catch(Exception $e)
-        {
-            dd($e->getMessage());
+        } catch (Exception $e) {
+            DB::rollBack();
+            return redirect()->back()->withErrors(['message' => 'Error: ' . $e->getMessage()]);
         }
     }
-
-    
 }
